backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.0
python-escpos==0.0.0
qrcode==7.4.2
requests==2.31.0
